xwayland: move geometry handling down the hierarchy
When transactions happen, the toplevels and unmanaged views will have
different handling for it.
