+++ src/view/xwayland/xwayland-toplevel-view.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "config.h"
+#include "wayfire/geometry.hpp"
 #include "xwayland-view-base.hpp"
 
 #if WF_HAS_XWAYLAND
@@ -10,6 +11,160 @@ class wayfire_xwayland_view : public wayfire_xwayland_view_base
     wf::wl_listener_wrapper on_request_move, on_request_resize,
         on_request_maximize, on_request_minimize, on_request_activate,
         on_request_fullscreen, on_set_parent, on_set_hints;
+
+    /**
+     * The bounding box of the view the last time it was rendered.
+     *
+     * This is used to damage the view when it is resized, because when a
+     * transformer changes because the view is resized, we can't reliably
+     * calculate the old view region to damage.
+     */
+    wf::geometry_t last_bounding_box{0, 0, 0, 0};
+
+    /** The output geometry of the view */
+    wf::geometry_t geometry{100, 100, 0, 0};
+
+    wf::geometry_t get_output_geometry() override
+    {
+        return geometry;
+    }
+
+    wf::geometry_t get_wm_geometry() override
+    {
+        if (priv->frame)
+        {
+            return priv->frame->expand_wm_geometry(geometry);
+        } else
+        {
+            return geometry;
+        }
+    }
+
+    void adjust_anchored_edge(wf::dimensions_t new_size)
+    {
+        if (priv->edges)
+        {
+            auto wm = get_wm_geometry();
+            if (priv->edges & WLR_EDGE_LEFT)
+            {
+                wm.x += geometry.width - new_size.width;
+            }
+
+            if (priv->edges & WLR_EDGE_TOP)
+            {
+                wm.y += geometry.height - new_size.height;
+            }
+
+            set_position(wm.x, wm.y, get_wm_geometry(), false);
+        }
+    }
+
+    void set_position(int x, int y, wf::geometry_t old_geometry, bool send_signal)
+    {
+        auto obox = get_output_geometry();
+        auto wm   = get_wm_geometry();
+
+        wf::view_geometry_changed_signal data;
+        data.view = self();
+        data.old_geometry = old_geometry;
+
+        view_damage_raw(self(), last_bounding_box);
+        /* obox.x - wm.x is the current difference in the output and wm geometry */
+        geometry.x = x + obox.x - wm.x;
+        geometry.y = y + obox.y - wm.y;
+
+        /* Make sure that if we move the view while it is unmapped, its snapshot
+         * is still valid coordinates */
+        priv->offscreen_buffer = priv->offscreen_buffer.translated({
+            x - data.old_geometry.x, y - data.old_geometry.y,
+        });
+
+        damage();
+
+        if (send_signal)
+        {
+            emit(&data);
+            wf::get_core().emit(&data);
+            if (get_output())
+            {
+                get_output()->emit(&data);
+            }
+        }
+
+        last_bounding_box = get_bounding_box();
+        wf::scene::update(this->get_surface_root_node(), wf::scene::update_flag::GEOMETRY);
+    }
+
+    void update_size()
+    {
+        if (!is_mapped())
+        {
+            return;
+        }
+
+        wf::dimensions_t current_size{
+            priv->wsurface->current.width,
+            priv->wsurface->current.height,
+        };
+
+        if ((current_size.width == geometry.width) &&
+            (current_size.height == geometry.height))
+        {
+            return;
+        }
+
+        /* Damage current size */
+        view_damage_raw(self(), last_bounding_box);
+        adjust_anchored_edge(current_size);
+
+        wf::view_geometry_changed_signal data;
+        data.view = self();
+        data.old_geometry = get_wm_geometry();
+
+        geometry.width  = current_size.width;
+        geometry.height = current_size.height;
+
+        /* Damage new size */
+        last_bounding_box = get_bounding_box();
+        view_damage_raw(self(), last_bounding_box);
+        emit(&data);
+        wf::get_core().emit(&data);
+        if (get_output())
+        {
+            get_output()->emit(&data);
+        }
+
+        wf::scene::update(this->get_surface_root_node(), wf::scene::update_flag::GEOMETRY);
+    }
+
+    void handle_client_configure(wlr_xwayland_surface_configure_event *ev) override
+    {
+        wf::point_t output_origin = {0, 0};
+        if (get_output())
+        {
+            output_origin = wf::origin(get_output()->get_layout_geometry());
+        }
+
+        if (!is_mapped())
+        {
+            /* If the view is not mapped yet, let it be configured as it
+             * wishes. We will position it properly in ::map() */
+            wlr_xwayland_surface_configure(xw, ev->x, ev->y, ev->width, ev->height);
+            if ((ev->mask & XCB_CONFIG_WINDOW_X) && (ev->mask & XCB_CONFIG_WINDOW_Y))
+            {
+                this->self_positioned = true;
+                this->geometry.x = ev->x - output_origin.x;
+                this->geometry.y = ev->y - output_origin.y;
+            }
+
+            return;
+        }
+
+        /* Use old x/y values */
+        ev->x = geometry.x + output_origin.x;
+        ev->y = geometry.y + output_origin.y;
+        configure_request(wlr_box{ev->x, ev->y, ev->width, ev->height});
+    }
 
   public:
     wayfire_xwayland_view(wlr_xwayland_surface *xww) :
@@ -136,6 +291,9 @@ class wayfire_xwayland_view : public wayfire_xwayland_view_base
         priv->keyboard_focus_enabled =
             wlr_xwayland_or_surface_wants_focus(xw);
 
+        geometry.width  = surface->current.width;
+        geometry.height = surface->current.height;
+
         if (xw->maximized_horz && xw->maximized_vert)
         {
             if ((xw->width > 0) && (xw->height > 0))
@@ -181,6 +339,19 @@ class wayfire_xwayland_view : public wayfire_xwayland_view_base
         /* Avoid loops where the client wants to have a certain size but the
          * compositor keeps trying to resize it */
         last_size_request = wf::dimensions(geometry);
+
+        update_size();
+
+        /* Clear the resize edges.
+         * This is must be done here because if the user(or plugin) resizes too fast,
+         * the shell client might still haven't configured the surface, and in this
+         * case the next commit(here) needs to still have access to the gravity */
+        if (!priv->in_continuous_resize)
+        {
+            priv->edges = 0;
+        }
+
+        this->last_bounding_box = get_bounding_box();
     }
 
     void set_moving(bool moving) override
@@ -193,6 +364,21 @@ class wayfire_xwayland_view : public wayfire_xwayland_view_base
         {
             send_configure();
         }
+    }
+
+    void move(int x, int y) override
+    {
+        set_position(x, y, get_wm_geometry(), true);
+        if (!priv->in_continuous_move)
+        {
+            send_configure();
+        }
+    }
+
+    void set_geometry(wf::geometry_t geometry) override
+    {
+        move(geometry.x, geometry.y);
+        resize(geometry.width, geometry.height);
     }
 
     void resize(int w, int h) override
