+++ src/view/xwayland/xwayland-unmanaged-view.hpp
@@ -12,6 +12,141 @@ class wayfire_unmanaged_xwayland_view : public wayfire_xwayland_view_base
   protected:
     wf::wl_listener_wrapper on_set_geometry;
 
+    /**
+     * The bounding box of the view the last time it was rendered.
+     *
+     * This is used to damage the view when it is resized, because when a
+     * transformer changes because the view is resized, we can't reliably
+     * calculate the old view region to damage.
+     */
+    wf::geometry_t last_bounding_box{0, 0, 0, 0};
+
+    /** The output geometry of the view */
+    wf::geometry_t geometry{100, 100, 0, 0};
+
+    wf::geometry_t get_output_geometry() override
+    {
+        return geometry;
+    }
+
+    wf::geometry_t get_wm_geometry() override
+    {
+        if (priv->frame)
+        {
+            return priv->frame->expand_wm_geometry(geometry);
+        } else
+        {
+            return geometry;
+        }
+    }
+
+    void set_position(int x, int y, wf::geometry_t old_geometry, bool send_signal)
+    {
+        wf::view_geometry_changed_signal data;
+        data.view = self();
+        data.old_geometry = old_geometry;
+
+        view_damage_raw(self(), last_bounding_box);
+        geometry.x = x;
+        geometry.y = y;
+
+        /* Make sure that if we move the view while it is unmapped, its snapshot
+         * is still valid coordinates */
+        priv->offscreen_buffer = priv->offscreen_buffer.translated({
+            x - data.old_geometry.x, y - data.old_geometry.y,
+        });
+
+        damage();
+
+        if (send_signal)
+        {
+            emit(&data);
+            wf::get_core().emit(&data);
+            if (get_output())
+            {
+                get_output()->emit(&data);
+            }
+        }
+
+        last_bounding_box = get_bounding_box();
+        wf::scene::update(this->get_surface_root_node(), wf::scene::update_flag::GEOMETRY);
+    }
+
+    void handle_client_configure(wlr_xwayland_surface_configure_event *ev) override
+    {
+        wf::point_t output_origin = {0, 0};
+        if (get_output())
+        {
+            output_origin = wf::origin(get_output()->get_layout_geometry());
+        }
+
+        if ((ev->mask & XCB_CONFIG_WINDOW_X) && (ev->mask & XCB_CONFIG_WINDOW_Y))
+        {
+            this->self_positioned = true;
+        } else
+        {
+            /* Use old x/y values */
+            ev->x = geometry.x + output_origin.x;
+            ev->y = geometry.y + output_origin.y;
+        }
+
+        if (!is_mapped())
+        {
+            /* If the view is not mapped yet, let it be configured as it
+             * wishes. We will position it properly in ::map() */
+            wlr_xwayland_surface_configure(xw, ev->x, ev->y, ev->width, ev->height);
+            if ((ev->mask & XCB_CONFIG_WINDOW_X) && (ev->mask & XCB_CONFIG_WINDOW_Y))
+            {
+                this->geometry.x = ev->x - output_origin.x;
+                this->geometry.y = ev->y - output_origin.y;
+            }
+        } else
+        {
+            configure_request({ev->x, ev->y, ev->width, ev->height});
+        }
+    }
+
+    void update_size()
+    {
+        if (!is_mapped())
+        {
+            return;
+        }
+
+        wf::dimensions_t current_size{
+            priv->wsurface->current.width,
+            priv->wsurface->current.height,
+        };
+
+        if ((current_size.width == geometry.width) &&
+            (current_size.height == geometry.height))
+        {
+            return;
+        }
+
+        /* Damage current size */
+        view_damage_raw(self(), last_bounding_box);
+
+        wf::view_geometry_changed_signal data;
+        data.view = self();
+        data.old_geometry = get_wm_geometry();
+
+        geometry.width  = current_size.width;
+        geometry.height = current_size.height;
+
+        /* Damage new size */
+        last_bounding_box = get_bounding_box();
+        view_damage_raw(self(), last_bounding_box);
+        emit(&data);
+        wf::get_core().emit(&data);
+        if (get_output())
+        {
+            get_output()->emit(&data);
+        }
+
+        wf::scene::update(this->get_surface_root_node(), wf::scene::update_flag::GEOMETRY);
+    }
+
   public:
     wayfire_unmanaged_xwayland_view(wlr_xwayland_surface *xww) : wayfire_xwayland_view_base(xww)
     {
@@ -37,7 +172,7 @@ class wayfire_unmanaged_xwayland_view : public wayfire_xwayland_view_base
                     geometry.y -= real_output.y;
                 }
 
-                wayfire_xwayland_view_base::move(geometry.x, geometry.y);
+                move(geometry.x, geometry.y);
             }
         });
 
@@ -74,7 +209,7 @@ class wayfire_unmanaged_xwayland_view : public wayfire_xwayland_view_base
 
         global_x = xw->x;
         global_y = xw->y;
-        wayfire_xwayland_view_base::move(xw->x - real_output_geometry.x, xw->y - real_output_geometry.y);
+        move(xw->x - real_output_geometry.x, xw->y - real_output_geometry.y);
 
         if (wo != get_output())
         {
@@ -104,6 +239,23 @@ class wayfire_unmanaged_xwayland_view : public wayfire_xwayland_view_base
         }
     }
 
+    void commit() override
+    {
+        wayfire_xwayland_view_base::commit();
+        update_size();
+
+        /* Clear the resize edges.
+         * This is must be done here because if the user(or plugin) resizes too fast,
+         * the shell client might still haven't configured the surface, and in this
+         * case the next commit(here) needs to still have access to the gravity */
+        if (!priv->in_continuous_resize)
+        {
+            priv->edges = 0;
+        }
+
+        this->last_bounding_box = get_bounding_box();
+    }
+
     void destroy() override
     {
         on_set_geometry.disconnect();
@@ -113,6 +265,38 @@ class wayfire_unmanaged_xwayland_view : public wayfire_xwayland_view_base
     bool should_be_decorated() override
     {
         return (!xw->override_redirect && !this->has_client_decoration);
+    }
+
+    void move(int x, int y) override
+    {
+        set_position(x, y, get_wm_geometry(), true);
+    }
+
+    void resize(int w, int h) override
+    {
+        if (priv->frame)
+        {
+            priv->frame->calculate_resize_size(w, h);
+        }
+
+        wf::dimensions_t current_size = {
+            get_output_geometry().width,
+            get_output_geometry().height
+        };
+
+        if (!should_resize_client({w, h}, current_size))
+        {
+            return;
+        }
+
+        this->last_size_request = {w, h};
+        send_configure(w, h);
+    }
+
+    void set_geometry(wf::geometry_t geometry) override
+    {
+        move(geometry.x, geometry.y);
+        resize(geometry.width, geometry.height);
     }
 
     wf::xw::view_type get_current_impl_type() const override
